fix: house fetchDisplayRate in a hook-free module — utils/currency pulls useCurrency into the server route
Vercel's webpack build rejected the API route importing utils/currency:
it re-exports through hooks/useCurrency (useState/useEffect), which is
client-only, and an RSC module may not depend on it. The local build had
passed on a stale .next cache. fx.utils.ts imports only actions/currency,
so both the server route and the Capacitor client branch can share it.
